# api/api-common/generated/openapi/src/main/java/com/diacore/api/operation/ProfileQueryApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.diacore.api.operation;

import com.diacore.api.model.CarbRatioListResponse;
import com.diacore.api.model.InsulinSensitivityListResponse;
import com.diacore.api.model.RecommendationResponse;
import com.diacore.api.model.TherapyParameterResponse;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-23T17:25:14.259910+09:00[Asia/Seoul]", comments = "Generator version: 7.5.0")
public interface ProfileQueryApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/v1/profiles/carb-ratios : get carb ratios (시간대별 탄수비 조회)
     *
     * @return success (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/profiles/carb-ratios",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CarbRatioListResponse> getCarbRatios(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"segments\" : [ { \"startTime\" : 8, \"value\" : 12.5 }, { \"startTime\" : 8, \"value\" : 12.5 } ], \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/profiles/insulin-sensitivities : get insulin sensitivities
     *
     * @return success (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/profiles/insulin-sensitivities",
        produces = { "application/json" }
    )
    
    default ResponseEntity<InsulinSensitivityListResponse> getInsulinSensitivities(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"segments\" : [ { \"startTime\" : 0, \"value\" : 50.0 }, { \"startTime\" : 0, \"value\" : 50.0 } ], \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/profiles/recommendations : get therapy parameters recommendation history
     *
     * @param userId  (required)
     * @return success (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/profiles/recommendations",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<RecommendationResponse>> getRecommendationHistory(
         @RequestParam(value = "userId", required = true) Long userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"parameterType\" : \"CARB_RATIO\", \"reasonText\" : \"최근 7일간 아침 식후 고혈당 지속\", \"originalValue\" : 6.0274563, \"id\" : 0, \"recommendedValue\" : 1.4658129 }, { \"parameterType\" : \"CARB_RATIO\", \"reasonText\" : \"최근 7일간 아침 식후 고혈당 지속\", \"originalValue\" : 6.0274563, \"id\" : 0, \"recommendedValue\" : 1.4658129 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/profiles/parameters : get active therapy parameters by hour
     *
     * @param hour 조회할 기준 시간 (0~23) (required)
     * @return success (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/profiles/parameters",
        produces = { "application/json" }
    )
    
    default ResponseEntity<TherapyParameterResponse> getTherapyParametersByHour(
         @RequestParam(value = "hour", required = true) Integer hour
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"hour\" : 15, \"carbRatio\" : 12.5, \"insulinSensitivity\" : 45.0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}