# api/api-common/generated/openapi/src/main/java/com/diacore/api/operation/SimulationQueryApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.diacore.api.operation;

import com.diacore.api.model.BolusRequest;
import com.diacore.api.model.ExerciseRequest;
import com.diacore.api.model.SimulationResponse;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-23T17:25:14.259910+09:00[Asia/Seoul]", comments = "Generator version: 7.5.0")
public interface SimulationQueryApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/v1/simulations/bolus : calculate bolus
     *
     * @param bolusRequest  (required)
     * @return success (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/simulations/bolus",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<SimulationResponse> calculateBolus(
         @RequestBody BolusRequest bolusRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"recommendedCarbsG\" : 6, \"predictedBg\" : 0, \"recommendedBolus\" : 1.4658129 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/simulations/exercise : predict exercise impact
     *
     * @param exerciseRequest  (required)
     * @return success (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/simulations/exercise",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<SimulationResponse> predictExerciseImpact(
         @RequestBody ExerciseRequest exerciseRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"recommendedCarbsG\" : 6, \"predictedBg\" : 0, \"recommendedBolus\" : 1.4658129 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}