# api/api-common/src/main/resources/openapi/diacore-api.yaml
openapi: 3.0.3
info:
  title: DiaCore Integration API
  version: 1.0.0
  description: 1형 당뇨 환자를 위한 데이터 기반 통합 관리 및 예측 플랫폼 DiaCore API 명세서

paths:
  # ==========================================
  # Log
  # ==========================================
  /api/v1/logs/glucose:
    post:
      tags:
        - LogCommand
      summary: add glucose log
      operationId: createGlucoseLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlucoseLogRequest'
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedResponse'
    get:
      tags:
        - LogQuery
      summary: get glucose logs
      operationId: getGlucoseLogs
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlucoseLogPageResponse'

  /api/v1/logs/meals:
    post:
      tags:
        - LogCommand
      summary: add meal log
      operationId: createMealLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealLogRequest'
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedResponse'
    get:
      tags:
        - LogQuery
      summary: get meal logs
      operationId: getMealLogs
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealLogPageResponse'

  /api/v1/logs/insulin:
    post:
      tags:
        - LogCommand
      summary: add insulin log
      operationId: createInsulinLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsulinLogRequest'
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedResponse'
    get:
      tags:
        - LogQuery
      summary: get insulin logs
      operationId: getInsulinLogs
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsulinLogPageResponse'

  # ==========================================
  # Profile
  # ==========================================
  /api/v1/profiles/carb-ratios:
    get:
      tags:
        - ProfileQuery
      summary: get carb ratios (시간대별 탄수비 조회)
      operationId: getCarbRatios
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarbRatioListResponse'
    put:
      tags:
        - ProfileCommand
      summary: update carb ratios (시간대별 탄수비 일괄 수정)
      operationId: updateCarbRatios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarbRatioListRequest'
      responses:
        '204':
          description: successfully updated

  /api/v1/profiles/insulin-sensitivities:
    get:
      tags:
        - ProfileQuery
      summary: get insulin sensitivities
      operationId: getInsulinSensitivities
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsulinSensitivityListResponse'
    put:
      tags:
        - ProfileCommand
      summary: update insulin sensitivities
      operationId: updateInsulinSensitivities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsulinSensitivityListRequest'
      responses:
        '204':
          description: successfully updated

  /api/v1/profiles/parameters:
    get:
      tags:
        - ProfileQuery
      summary: get active therapy parameters by hour
      operationId: getTherapyParametersByHour
      parameters:
        - name: hour
          in: query
          required: true
          description: "조회할 기준 시간 (0~23)"
          schema:
            type: integer
            minimum: 0
            maximum: 23
            example: 15
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TherapyParameterResponse'

  /api/v1/profiles/recommendations:
    get:
      tags:
        - ProfileQuery
      summary: get therapy parameters recommendation history
      operationId: getRecommendationHistory
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecommendationResponse'

  # ==========================================
  # Simulation
  # ==========================================
  /api/v1/simulations/exercise:
    post:
      tags:
        - SimulationQuery
      summary: predict exercise impact
      operationId: predictExerciseImpact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'

  /api/v1/simulations/bolus:
    post:
      tags:
        - SimulationQuery
      summary: calculate bolus
      operationId: calculateBolus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BolusRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'

  # ==========================================
  # User
  # ==========================================
  /api/v1/users/register:
    post:
      summary: Register User
      operationId: registerUser
      tags:
        - User Command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedResponse'

  /api/v1/users/login:
    post:
      summary: Login
      operationId: loginUser
      tags:
        - User Query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  /api/v1/users/withdraw:
    delete:
      summary: Withdraw User
      operationId: withdrawUser
      tags:
        - User Command
      responses:
        '204':
          description: success

components:
  schemas:
    CommonResponse:
      type: object
      required:
        - success
        - code
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: API 요청 성공 여부
          example: true
        code:
          type: string
          description: 비즈니스 응답 코드
          example: "SUCCESS"
        message:
          type: string
          description: 응답 메시지
          example: "요청이 성공적으로 처리되었습니다."
        data:
          type: object
          description: 응답 데이터
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: 응답 발생 시각

    CreatedResponse:
      type: object
      required: [ id, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: 생성된 리소스의 ID
        createdAt:
          type: string
          format: date-time
          description: 생성된 시각

    GlucoseLogRequest:
      type: object
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: integer
          description: "glucose (mg/dL)"
        trendArrow:
          type: string
          example: "FLAT"
        source:
          type: string
          example: "dexcom G7"

    MealLogRequest:
      type: object
      required: [timestamp, carbohydrateG, mealType]
      properties:
        timestamp:
          type: string
          format: date-time
        carbohydrateG:
          type: float
          description: "carb (g)"
        mealType:
          type: string
          example: "BREAKFAST"
        foodDesc:
          type: string
          example: "사과 1개"

    InsulinLogRequest:
      type: object
      required: [timestamp, dose, insulinType]
      properties:
        timestamp:
          type: string
          format: date-time
        dose:
          type: number
          format: float
          description: "insulin (Unit)"
        insulinType:
          type: string
          example: "RAPID_ACTING"

    TherapyParametersResponse:
      type: object
      properties:
        carbRatio:
          type: number
          format: float
          description: "carb ratio (g/U)"
        insulinSensitivity:
          type: number
          format: float
          description: "insulin sensitivity (mg/dL/U)"

    TherapyParametersRequest:
      type: object
      required: [carbRatio, insulinSensitivity]
      properties:
        carbRatio:
          type: number
          format: float
        insulinSensitivity:
          type: number
          format: float

    RecommendationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        parameterType:
          type: string
          example: "CARB_RATIO"
        originalValue:
          type: number
          format: float
        recommendedValue:
          type: number
          format: float
        reasonText:
          type: string
          example: "최근 7일간 아침 식후 고혈당 지속"

    ExerciseRequest:
      type: object
      required: [exerciseType, durationMinutes, currentBg]
      properties:
        exerciseType:
          type: string
          example: "RUNNING"
        durationMinutes:
          type: integer
        currentBg:
          type: integer

    BolusRequest:
      type: object
      required: [currentBg, targetBg, mealCarbsG]
      properties:
        currentBg:
          type: integer
        targetBg:
          type: integer
        mealCarbsG:
          type: integer

    SimulationResponse:
      type: object
      properties:
        predictedBg:
          type: integer
          description: "result glucose"
        recommendedCarbsG:
          type: integer
          description: "recommended carb (g)"
        recommendedBolus:
          type: number
          format: float
          description: "recommended insulin (Unit)"

    RegisterUserRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "securePassword123!"
        name:
          type: string
          example: "사용자"

    LoginUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "securePassword123!"

    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Access Token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    GlucoseLogResponse:
      type: object
      required: [ id, timestamp, value ]
      properties:
        id: { type: integer, format: int64 }
        timestamp: { type: string, format: date-time }
        value: { type: integer }
        trendArrow: { type: string }
        source: { type: string }

    MealLogResponse:
      type: object
      required: [ id, timestamp, carbohydrateG, mealType ]
      properties:
        id: { type: integer, format: int64 }
        timestamp: { type: string, format: date-time }
        carbohydrateG: { type: float }
        mealType: { type: string }
        foodDesc: { type: string }

    InsulinLogResponse:
      type: object
      required: [ id, timestamp, dose, insulinType ]
      properties:
        id: { type: integer, format: int64 }
        timestamp: { type: string, format: date-time }
        dose: { type: number, format: float }
        insulinType: { type: string }

    GlucoseLogPageResponse:
      type: object
      required: [ content, totalElements ]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/GlucoseLogResponse'
        totalElements: { type: integer, format: int64 }
        pageNumber: { type: integer }

    MealLogPageResponse:
      type: object
      required: [ content, totalElements ]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MealLogResponse'
        totalElements: { type: integer, format: int64 }
        pageNumber: { type: integer }

    InsulinLogPageResponse:
      type: object
      required: [ content, totalElements ]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InsulinLogResponse'
        totalElements: { type: integer, format: int64 }
        pageNumber: { type: integer }

    CarbRatioSegment:
      type: object
      required: [ startTime, value ]
      properties:
        startTime:
          type: integer
          description: "시작 시간 (0~23)"
          example: 8
        value:
          type: number
          format: float
          description: "탄수비 (g/U)"
          example: 12.5

    CarbRatioListRequest:
      type: object
      required: [ segments ]
      properties:
        segments:
          type: array
          items:
            $ref: '#/components/schemas/CarbRatioSegment'

    CarbRatioListResponse:
      type: object
      required: [ segments, updatedAt ]
      properties:
        segments:
          type: array
          items:
            $ref: '#/components/schemas/CarbRatioSegment'
        updatedAt:
          type: string
          format: date-time

    InsulinSensitivitySegment:
      type: object
      required: [ startTime, value ]
      properties:
        startTime:
          type: integer
          description: "시작 시간 (0~23)"
          example: 0
        value:
          type: number
          format: float
          description: "인슐린 민감도 (mg/dL/U)"
          example: 50.0

    InsulinSensitivityListRequest:
      type: object
      required: [ segments ]
      properties:
        segments:
          type: array
          items:
            $ref: '#/components/schemas/InsulinSensitivitySegment'

    InsulinSensitivityListResponse:
      type: object
      required: [ segments, updatedAt ]
      properties:
        segments:
          type: array
          items:
            $ref: '#/components/schemas/InsulinSensitivitySegment'
        updatedAt:
          type: string
          format: date-time

    TherapyParameterResponse:
      type: object
      required: [ hour, carbRatio, insulinSensitivity ]
      properties:
        hour:
          type: integer
          description: "요청한 기준 시간"
          example: 15
        carbRatio:
          type: number
          format: float
          description: "해당 시간에 적용되는 탄수비 (g/U)"
          example: 12.5
        insulinSensitivity:
          type: number
          format: float
          description: "해당 시간에 적용되는 인슐린 민감도 (mg/dL/U)"
          example: 45.0